+++ src/plone/mls/listing/portlets/agent_contact.py
@@ -4,7 +4,6 @@
 # python imports
 from email import message_from_string
 import copy
-import formencode
 import logging
 import re
 
@@ -13,7 +12,6 @@ from Acquisition import aq_inner
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import PloneMessageFactory as PMF
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from plone import api as ploneapi
 from plone.app.form.widgets.uberselectionwidget import UberSelectionWidget
 from plone.app.portlets.portlets import base
 from plone.app.vocabularies.catalog import SearchableTextSourceBinder
@@ -95,33 +93,6 @@ def validate_email(value):
     return True
 
 
-def update_formencode_i18n():
-    try:
-        language = ploneapi.portal.get_current_language()
-    except AttributeError:
-        logger.warning('plone.api version is too old.')
-        return
-    try:
-        formencode.api.set_stdtranslation(
-            domain='FormEncode',
-            languages=[language],
-        )
-    except Exception, e:
-        logger.warning(e)
-
-
-def validate_international_phone(value):
-    """Check for valid international phone numbers."""
-    if value:
-        update_formencode_i18n()
-        c = formencode.national.InternationalPhoneNumber()
-        try:
-            c.to_python(value)
-        except (formencode.api.Invalid), error:
-            raise Invalid(error)
-    return True
-
-
 def contains_nuts(value):
     """Check for traces of nuts, like urls or other spammer fun things"""
     if value:
@@ -170,7 +141,6 @@ class IEmailForm(Interface):
     )
 
     phone = schema.TextLine(
-        constraint=validate_international_phone,
         description=_(
             u'Please enter a phone number. Some agents will not respond '
             u'without one.'